requirements.txt
streamlit
streamlit-chat
langchain
openai
python-dotenv

